Move RequestIdGuard to types::client
